test(cart): split 1619-line test_repositories.py by class group (#109, A-8)
Splits services/cart/tests/unit/test_repositories.py (1619 lines, 16
test classes) into four focused files. Each new file is under ~810
lines, sharing the same header of imports / helper functions
(_make_terminal_info, _make_mock_db, _make_aiohttp_context_manager).

  test_cart_repository.py            (707 lines)  — Cart-specific:
    CircuitBreaker, ShardKey, CreateCart, CacheCartAsync,
    DaprCacheAsync, DaprGetCachedAsync, DaprDeleteCachedAsync,
    DBFallbackSave, DBFallbackGet, DBFallbackDelete,
    DeleteCartAsync.

  test_transaction_repositories.py   (810 lines)  — Transaction-side:
    TransactionStatusRepository, TranlogDeliveryStatusRepository,
    TranlogRepository.

  test_terminal_counter_repository.py (178 lines) — TerminalCounter.

  test_tax_master_repository.py      (183 lines) — TaxMaster.

Verified: pipenv run pytest -m unit → 452 passed (same as before).

Refs #109, follows up #113 — the other large files
(master-data/test_repositories.py 1424 lines,
terminal/test_terminal_service.py 1154 lines, etc.) remain in #113
for follow-up.
