+++ services/cart/tests/unit/test_transaction_repositories.py
@@ -1,0 +1,810 @@
+"""Unit tests split out from test_repositories.py.
+
+Shared helpers / imports preserved as-is from the original
+test_repositories.py — splitting by repository class group keeps
+each file under ~700 lines and lets pytest-xdist parallelise faster.
+"""
+import time
+import sys
+from datetime import datetime, timedelta
+from unittest.mock import AsyncMock, MagicMock, patch, PropertyMock
+
+import pytest
+import pytest_asyncio
+
+from kugel_common.models.documents.terminal_info_document import TerminalInfoDocument
+from kugel_common.models.documents.staff_master_document import StaffMasterDocument
+from kugel_common.models.documents.user_info_document import UserInfoDocument
+from kugel_common.exceptions import (
+    NotFoundException,
+    LoadDataNoExistException,
+    CannotCreateException,
+)
+
+from app.enums.cart_status import CartStatus
+from app.models.documents.cart_document import CartDocument
+from app.models.documents.tax_master_document import TaxMasterDocument
+from app.models.documents.settings_master_document import SettingsMasterDocument
+from app.models.documents.item_master_document import ItemMasterDocument
+from app.models.documents.transaction_status_document import TransactionStatusDocument
+from app.models.documents.tranlog_delivery_status_document import TranlogDeliveryStatus
+from app.models.documents.terminal_counter_document import TerminalCounterDocument
+
+from app.models.repositories.cart_repository import CartRepository
+from app.models.repositories.transaction_status_repository import TransactionStatusRepository
+from app.models.repositories.tranlog_delivery_status_repository import TranlogDeliveryStatusRepository
+from app.models.repositories.tranlog_repository import TranlogRepository
+from app.models.repositories.terminal_counter_repository import (
+    TerminalCounterRepository,
+    make_terminal_id,
+)
+from app.models.repositories.tax_master_repository import TaxMasterRepository
+
+
+# ---------------------------------------------------------------------------
+# Fixtures
+# ---------------------------------------------------------------------------
+
+def _make_terminal_info(**overrides) -> TerminalInfoDocument:
+    """Create a TerminalInfoDocument with sensible defaults."""
+    defaults = dict(
+        tenant_id="T001",
+        store_code="S001",
+        terminal_no=1,
+        business_date="20240601",
+        open_counter=1,
+        staff=StaffMasterDocument(id="staff01", name="Test Staff"),
+    )
+    defaults.update(overrides)
+    return TerminalInfoDocument(**defaults)
+
+
+def _make_mock_db():
+    """Return a MagicMock that behaves like AsyncIOMotorDatabase."""
+    db = MagicMock()
+    db.get_collection = MagicMock(return_value=MagicMock())
+    db.client = MagicMock()
+    return db
+
+
+# =========================================================================
+# CartRepository tests
+# =========================================================================
+
+def _make_aiohttp_context_manager(mock_response):
+    """Wrap a mock response to behave as an aiohttp async context manager.
+
+    aiohttp session methods (post/get/delete) return an object that supports
+    ``async with`` directly (not a coroutine returning a context manager).
+    Using MagicMock ensures the return value is synchronous while still
+    supporting ``__aenter__``/``__aexit__``.
+    """
+    cm = MagicMock()
+    cm.__aenter__ = AsyncMock(return_value=mock_response)
+    cm.__aexit__ = AsyncMock(return_value=False)
+    return cm
+
+
+
+class TestTransactionStatusRepository:
+    """Tests for TransactionStatusRepository business logic."""
+
+    def _make_repo(self, db=None):
+        db = db or _make_mock_db()
+        terminal_info = _make_terminal_info()
+        repo = TransactionStatusRepository(db, terminal_info)
+        return repo
+
+    @pytest.mark.asyncio
+    async def test_mark_as_voided_creates_new_when_no_existing(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+        repo.create_async = AsyncMock(return_value=True)
+
+        result = await repo.mark_as_voided_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            void_transaction_no=101,
+            staff_id="staff01",
+        )
+
+        assert result.is_voided is True
+        assert result.void_transaction_no == 101
+        assert result.void_staff_id == "staff01"
+        assert result.tenant_id == "T001"
+        assert result.transaction_no == 100
+        repo.create_async.assert_awaited_once()
+
+    @pytest.mark.asyncio
+    async def test_mark_as_voided_preserves_refund_info_on_existing(self):
+        repo = self._make_repo()
+
+        existing = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_refunded=True,
+            return_transaction_no=50,
+            return_date_time="2024-06-01T10:00:00",
+            return_staff_id="staff_ret",
+        )
+
+        # First call returns existing, second call (after update) returns updated
+        updated = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_voided=True,
+            is_refunded=True,
+            void_transaction_no=101,
+            return_transaction_no=50,
+        )
+        repo.get_one_async = AsyncMock(side_effect=[existing, updated])
+        repo.update_one_async = AsyncMock(return_value=True)
+
+        result = await repo.mark_as_voided_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            void_transaction_no=101,
+            staff_id="staff01",
+        )
+
+        # Verify update was called with preserved refund fields
+        update_call_args = repo.update_one_async.call_args
+        update_values = update_call_args[0][1]
+        assert update_values["is_voided"] is True
+        assert update_values["is_refunded"] is True
+        assert update_values["return_transaction_no"] == 50
+        assert update_values["return_date_time"] == "2024-06-01T10:00:00"
+        assert update_values["return_staff_id"] == "staff_ret"
+
+    @pytest.mark.asyncio
+    async def test_mark_as_refunded_creates_new_when_no_existing(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+        repo.create_async = AsyncMock(return_value=True)
+
+        result = await repo.mark_as_refunded_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            return_transaction_no=102,
+            staff_id="staff02",
+        )
+
+        assert result.is_refunded is True
+        assert result.return_transaction_no == 102
+        assert result.return_staff_id == "staff02"
+        repo.create_async.assert_awaited_once()
+
+    @pytest.mark.asyncio
+    async def test_mark_as_refunded_preserves_void_info_on_existing(self):
+        repo = self._make_repo()
+
+        existing = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_voided=True,
+            void_transaction_no=90,
+            void_date_time="2024-05-01T09:00:00",
+            void_staff_id="staff_void",
+        )
+
+        updated = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_voided=True,
+            is_refunded=True,
+        )
+        repo.get_one_async = AsyncMock(side_effect=[existing, updated])
+        repo.update_one_async = AsyncMock(return_value=True)
+
+        result = await repo.mark_as_refunded_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            return_transaction_no=102,
+            staff_id="staff02",
+        )
+
+        update_call_args = repo.update_one_async.call_args
+        update_values = update_call_args[0][1]
+        assert update_values["is_refunded"] is True
+        assert update_values["is_voided"] is True
+        assert update_values["void_transaction_no"] == 90
+        assert update_values["void_date_time"] == "2024-05-01T09:00:00"
+        assert update_values["void_staff_id"] == "staff_void"
+
+    @pytest.mark.asyncio
+    async def test_get_status_for_transactions_builds_in_query(self):
+        repo = self._make_repo()
+
+        status1 = TransactionStatusDocument(
+            tenant_id="T001", store_code="S001", terminal_no=1, transaction_no=10
+        )
+        status2 = TransactionStatusDocument(
+            tenant_id="T001", store_code="S001", terminal_no=1, transaction_no=20
+        )
+        repo.get_list_async = AsyncMock(return_value=[status1, status2])
+
+        result = await repo.get_status_for_transactions_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_nos=[10, 20, 30],
+        )
+
+        # Verify query contains $in operator
+        query = repo.get_list_async.call_args[0][0]
+        assert query["transaction_no"] == {"$in": [10, 20, 30]}
+        assert query["tenant_id"] == "T001"
+        assert query["store_code"] == "S001"
+        assert query["terminal_no"] == 1
+
+        # Verify result is a dictionary keyed by transaction_no
+        assert isinstance(result, dict)
+        assert 10 in result
+        assert 20 in result
+        assert 30 not in result
+
+    @pytest.mark.asyncio
+    async def test_get_status_for_transactions_returns_empty_dict_when_none_found(self):
+        repo = self._make_repo()
+        repo.get_list_async = AsyncMock(return_value=[])
+
+        result = await repo.get_status_for_transactions_async(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_nos=[99],
+        )
+
+        assert result == {}
+
+    @pytest.mark.asyncio
+    async def test_reset_refund_status_clears_refund_fields(self):
+        repo = self._make_repo()
+
+        existing = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_refunded=True,
+            return_transaction_no=50,
+        )
+        updated = TransactionStatusDocument(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            is_refunded=False,
+        )
+        repo.get_one_async = AsyncMock(side_effect=[existing, updated])
+        repo.update_one_async = AsyncMock(return_value=True)
+
+        result = await repo.reset_refund_status_async(
+            tenant_id="T001", store_code="S001", terminal_no=1, transaction_no=100
+        )
+
+        update_call_args = repo.update_one_async.call_args
+        update_values = update_call_args[0][1]
+        assert update_values["is_refunded"] is False
+        assert update_values["return_transaction_no"] is None
+        assert update_values["return_date_time"] is None
+        assert update_values["return_staff_id"] is None
+
+    @pytest.mark.asyncio
+    async def test_reset_refund_status_returns_none_when_no_existing(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+
+        result = await repo.reset_refund_status_async(
+            tenant_id="T001", store_code="S001", terminal_no=1, transaction_no=999
+        )
+
+        assert result is None
+
+    @pytest.mark.asyncio
+    async def test_get_status_by_transaction_builds_correct_query(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+
+        await repo.get_status_by_transaction_async("T001", "S001", 1, 100)
+
+        query = repo.get_one_async.call_args[0][0]
+        assert query == {
+            "tenant_id": "T001",
+            "store_code": "S001",
+            "terminal_no": 1,
+            "transaction_no": 100,
+        }
+
+
+# =========================================================================
+# TranlogDeliveryStatusRepository tests
+# =========================================================================
+
+
+class TestTranlogDeliveryStatusRepository:
+    """Tests for TranlogDeliveryStatusRepository."""
+
+    def _make_repo(self, db=None):
+        db = db or _make_mock_db()
+        terminal_info = _make_terminal_info()
+        repo = TranlogDeliveryStatusRepository(db, terminal_info)
+        return repo
+
+    @pytest.mark.asyncio
+    async def test_create_status_constructs_service_status_list(self):
+        repo = self._make_repo()
+        repo.create_async = AsyncMock(return_value=True)
+
+        services = [
+            {"service_name": "report", "status": "pending"},
+            {"service_name": "journal", "status": "pending"},
+        ]
+
+        result = await repo.create_status_async(
+            event_id="evt-001",
+            transaction_no=100,
+            payload={"data": "test"},
+            services=services,
+        )
+
+        assert result is True
+        created_doc = repo.create_async.call_args[0][0]
+        assert isinstance(created_doc, TranlogDeliveryStatus)
+        assert created_doc.event_id == "evt-001"
+        assert created_doc.transaction_no == 100
+        assert created_doc.tenant_id == "T001"
+        assert created_doc.store_code == "S001"
+        assert created_doc.terminal_no == 1
+        assert created_doc.business_date == "20240601"
+        assert created_doc.open_counter == 1
+        assert created_doc.status == "published"
+        assert len(created_doc.services) == 2
+        assert created_doc.services[0].service_name == "report"
+        assert created_doc.services[1].service_name == "journal"
+
+    @pytest.mark.asyncio
+    async def test_create_status_with_no_services(self):
+        repo = self._make_repo()
+        repo.create_async = AsyncMock(return_value=True)
+
+        await repo.create_status_async(
+            event_id="evt-002",
+            transaction_no=200,
+            payload={},
+            services=None,
+        )
+
+        created_doc = repo.create_async.call_args[0][0]
+        assert created_doc.services == []
+
+    @pytest.mark.asyncio
+    async def test_find_pending_deliveries_builds_correct_query(self):
+        repo = self._make_repo()
+        repo.get_list_async = AsyncMock(return_value=[])
+
+        await repo.find_pending_deliveries(hours_ago=12)
+
+        query = repo.get_list_async.call_args[0][0]
+        assert "published_at" in query
+        assert "$gte" in query["published_at"]
+        assert query["status"] == {"$nin": ["delivered"]}
+
+    @pytest.mark.asyncio
+    async def test_find_pending_deliveries_default_hours(self):
+        repo = self._make_repo()
+        repo.get_list_async = AsyncMock(return_value=[])
+
+        await repo.find_pending_deliveries()
+
+        query = repo.get_list_async.call_args[0][0]
+        threshold = query["published_at"]["$gte"]
+        assert isinstance(threshold, datetime)
+
+    @pytest.mark.asyncio
+    async def test_find_by_event_id_uses_correct_filter(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+
+        await repo.find_by_event_id("evt-xyz")
+
+        repo.get_one_async.assert_awaited_once_with({"event_id": "evt-xyz"})
+
+    @pytest.mark.asyncio
+    async def test_find_by_transaction_info_uses_correct_filter(self):
+        repo = self._make_repo()
+        repo.get_list_async = AsyncMock(return_value=[])
+
+        await repo.find_by_transaction_info("T001", "S001", 1, 100)
+
+        expected_filter = {
+            "tenant_id": "T001",
+            "store_code": "S001",
+            "terminal_no": 1,
+            "transaction_no": 100,
+        }
+        repo.get_list_async.assert_awaited_once_with(expected_filter)
+
+    @pytest.mark.asyncio
+    async def test_find_by_business_date_uses_correct_filter(self):
+        repo = self._make_repo()
+        repo.get_list_async = AsyncMock(return_value=[])
+
+        await repo.find_by_business_date("T001", "S001", "20240601")
+
+        expected_filter = {
+            "tenant_id": "T001",
+            "store_code": "S001",
+            "business_date": "20240601",
+        }
+        repo.get_list_async.assert_awaited_once_with(expected_filter)
+
+    @pytest.mark.asyncio
+    async def test_update_service_status_builds_array_filter(self):
+        repo = self._make_repo()
+
+        # Set up mock collection
+        mock_collection = MagicMock()
+        mock_update_result = MagicMock()
+        mock_update_result.modified_count = 1
+        mock_collection.update_one = AsyncMock(return_value=mock_update_result)
+        repo.dbcollection = mock_collection
+
+        # Mock find_by_event_id to return a doc with all services delivered
+        delivered_doc = TranlogDeliveryStatus(
+            event_id="evt-001",
+            published_at=datetime.now(),
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            business_date="20240601",
+            open_counter=1,
+            payload={},
+            services=[
+                TranlogDeliveryStatus.ServiceStatus(service_name="report", status="delivered"),
+                TranlogDeliveryStatus.ServiceStatus(service_name="journal", status="delivered"),
+            ],
+            last_updated_at=datetime.now(),
+        )
+        repo.find_by_event_id = AsyncMock(return_value=delivered_doc)
+        repo.update_delivery_status = AsyncMock(return_value=True)
+
+        result = await repo.update_service_status(
+            event_id="evt-001",
+            service_name="report",
+            status="delivered",
+        )
+
+        assert result is True
+        # Verify array filter was passed
+        update_call = mock_collection.update_one.call_args
+        assert update_call.kwargs["array_filters"] == [{"elem.service_name": "report"}]
+        # Overall status should be set to delivered since all services are delivered
+        repo.update_delivery_status.assert_awaited_once_with("evt-001", "delivered")
+
+    @pytest.mark.asyncio
+    async def test_update_service_status_sets_partially_delivered_on_failure(self):
+        repo = self._make_repo()
+
+        mock_collection = MagicMock()
+        mock_update_result = MagicMock()
+        mock_update_result.modified_count = 1
+        mock_collection.update_one = AsyncMock(return_value=mock_update_result)
+        repo.dbcollection = mock_collection
+
+        failed_doc = TranlogDeliveryStatus(
+            event_id="evt-002",
+            published_at=datetime.now(),
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            business_date="20240601",
+            open_counter=1,
+            payload={},
+            services=[
+                TranlogDeliveryStatus.ServiceStatus(service_name="report", status="delivered"),
+                TranlogDeliveryStatus.ServiceStatus(service_name="journal", status="failed"),
+            ],
+            last_updated_at=datetime.now(),
+        )
+        repo.find_by_event_id = AsyncMock(return_value=failed_doc)
+        repo.update_delivery_status = AsyncMock(return_value=True)
+
+        await repo.update_service_status(
+            event_id="evt-002",
+            service_name="journal",
+            status="failed",
+        )
+
+        repo.update_delivery_status.assert_awaited_once_with("evt-002", "partially_delivered")
+
+    @pytest.mark.asyncio
+    async def test_update_service_status_sets_published_when_pending_exists(self):
+        repo = self._make_repo()
+
+        mock_collection = MagicMock()
+        mock_update_result = MagicMock()
+        mock_update_result.modified_count = 1
+        mock_collection.update_one = AsyncMock(return_value=mock_update_result)
+        repo.dbcollection = mock_collection
+
+        pending_doc = TranlogDeliveryStatus(
+            event_id="evt-003",
+            published_at=datetime.now(),
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=100,
+            business_date="20240601",
+            open_counter=1,
+            payload={},
+            services=[
+                TranlogDeliveryStatus.ServiceStatus(service_name="report", status="delivered"),
+                TranlogDeliveryStatus.ServiceStatus(service_name="journal", status="pending"),
+            ],
+            last_updated_at=datetime.now(),
+        )
+        repo.find_by_event_id = AsyncMock(return_value=pending_doc)
+        repo.update_delivery_status = AsyncMock(return_value=True)
+
+        await repo.update_service_status(
+            event_id="evt-003",
+            service_name="report",
+            status="delivered",
+        )
+
+        repo.update_delivery_status.assert_awaited_once_with("evt-003", "published")
+
+    @pytest.mark.asyncio
+    async def test_update_service_status_with_message(self):
+        repo = self._make_repo()
+
+        mock_collection = MagicMock()
+        mock_update_result = MagicMock()
+        mock_update_result.modified_count = 0
+        mock_collection.update_one = AsyncMock(return_value=mock_update_result)
+        repo.dbcollection = mock_collection
+
+        result = await repo.update_service_status(
+            event_id="evt-004",
+            service_name="journal",
+            status="failed",
+            message="Connection timeout",
+        )
+
+        assert result is False
+        # Verify that message was included in update dict
+        update_call = mock_collection.update_one.call_args
+        update_dict = update_call[0][1]
+        assert "services.$[elem].message" in update_dict["$set"]
+        assert update_dict["$set"]["services.$[elem].message"] == "Connection timeout"
+
+    @pytest.mark.asyncio
+    async def test_update_service_status_returns_false_on_exception(self):
+        repo = self._make_repo()
+
+        mock_collection = MagicMock()
+        mock_collection.update_one = AsyncMock(side_effect=Exception("DB error"))
+        repo.dbcollection = mock_collection
+
+        result = await repo.update_service_status(
+            event_id="evt-005",
+            service_name="report",
+            status="delivered",
+        )
+
+        assert result is False
+
+    @pytest.mark.asyncio
+    async def test_update_delivery_status_uses_correct_filter_and_retries(self):
+        repo = self._make_repo()
+        repo.update_one_async = AsyncMock(return_value=True)
+
+        result = await repo.update_delivery_status("evt-001", "delivered")
+
+        assert result is True
+        call_args = repo.update_one_async.call_args
+        assert call_args[0][0] == {"event_id": "evt-001"}
+        assert call_args[0][1]["status"] == "delivered"
+        assert call_args[1]["max_retries"] == 10
+
+
+# =========================================================================
+# TranlogRepository tests
+# =========================================================================
+
+
+class TestTranlogRepository:
+    """Tests for TranlogRepository query building and shard key generation."""
+
+    def _make_repo(self, db=None):
+        db = db or _make_mock_db()
+        terminal_info = _make_terminal_info()
+        repo = TranlogRepository(db, terminal_info)
+        return repo
+
+    @pytest.mark.asyncio
+    async def test_query_with_all_optional_params(self):
+        repo = self._make_repo()
+        repo.get_paginated_list_async = AsyncMock(return_value=MagicMock())
+
+        await repo.get_tranlog_list_by_query_async(
+            store_code="S001",
+            terminal_no=1,
+            business_date="20240601",
+            open_counter=5,
+            transaction_type=[1, 2],
+            receipt_no=100,
+            limit=50,
+            page=2,
+            sort=[("transaction_no", -1)],
+            include_cancelled=True,
+        )
+
+        call_args = repo.get_paginated_list_async.call_args
+        query = call_args[1]["filter"]
+        assert query["tenant_id"] == "T001"
+        assert query["store_code"] == "S001"
+        assert query["terminal_no"] == 1
+        assert query["business_date"] == "20240601"
+        assert query["open_counter"] == 5
+        assert query["transaction_type"] == {"$in": [1, 2]}
+        assert query["receipt_no"] == 100
+        assert "sales.is_cancelled" not in query  # include_cancelled=True
+        assert call_args[1]["limit"] == 50
+        assert call_args[1]["page"] == 2
+        assert call_args[1]["sort"] == [("transaction_no", -1)]
+
+    @pytest.mark.asyncio
+    async def test_query_with_no_optional_params(self):
+        repo = self._make_repo()
+        repo.get_paginated_list_async = AsyncMock(return_value=MagicMock())
+
+        await repo.get_tranlog_list_by_query_async(
+            store_code="S001",
+            terminal_no=1,
+        )
+
+        call_args = repo.get_paginated_list_async.call_args
+        query = call_args[1]["filter"]
+        assert query == {
+            "tenant_id": "T001",
+            "store_code": "S001",
+            "terminal_no": 1,
+            "sales.is_cancelled": False,
+        }
+
+    @pytest.mark.asyncio
+    async def test_query_excludes_cancelled_by_default(self):
+        repo = self._make_repo()
+        repo.get_paginated_list_async = AsyncMock(return_value=MagicMock())
+
+        await repo.get_tranlog_list_by_query_async(
+            store_code="S001",
+            terminal_no=1,
+        )
+
+        call_args = repo.get_paginated_list_async.call_args
+        query = call_args[1]["filter"]
+        assert query["sales.is_cancelled"] is False
+
+    @pytest.mark.asyncio
+    async def test_query_includes_cancelled_when_flag_set(self):
+        repo = self._make_repo()
+        repo.get_paginated_list_async = AsyncMock(return_value=MagicMock())
+
+        await repo.get_tranlog_list_by_query_async(
+            store_code="S001",
+            terminal_no=1,
+            include_cancelled=True,
+        )
+
+        call_args = repo.get_paginated_list_async.call_args
+        query = call_args[1]["filter"]
+        assert "sales.is_cancelled" not in query
+
+    def test_shard_key_generation(self):
+        repo = self._make_repo()
+        from kugel_common.models.documents.base_tranlog import BaseTransaction
+
+        tranlog = BaseTransaction(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            generate_date_time="2024-06-01T10:30:00Z",
+        )
+        key = repo._TranlogRepository__get_shard_key(tranlog)
+        assert key == "T001_S001_1_2024-06-01"
+
+    def test_shard_key_with_different_datetime_format(self):
+        repo = self._make_repo()
+        from kugel_common.models.documents.base_tranlog import BaseTransaction
+
+        tranlog = BaseTransaction(
+            tenant_id="TENX",
+            store_code="STR2",
+            terminal_no=99,
+            generate_date_time="2025-12-31T23:59:59Z",
+        )
+        key = repo._TranlogRepository__get_shard_key(tranlog)
+        assert key == "TENX_STR2_99_2025-12-31"
+
+    @pytest.mark.asyncio
+    async def test_create_tranlog_sets_shard_key_and_calls_create(self):
+        repo = self._make_repo()
+        repo.create_async = AsyncMock(return_value=True)
+
+        from kugel_common.models.documents.base_tranlog import BaseTransaction
+
+        tranlog = BaseTransaction(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=500,
+            transaction_type=1,
+            generate_date_time="2024-06-01T12:00:00Z",
+        )
+
+        result = await repo.create_tranlog_async(tranlog)
+
+        assert result.shard_key == "T001_S001_1_2024-06-01"
+        repo.create_async.assert_awaited_once()
+
+    @pytest.mark.asyncio
+    async def test_create_tranlog_raises_on_failure(self):
+        repo = self._make_repo()
+        repo.create_async = AsyncMock(return_value=False)
+
+        from kugel_common.models.documents.base_tranlog import BaseTransaction
+
+        tranlog = BaseTransaction(
+            tenant_id="T001",
+            store_code="S001",
+            terminal_no=1,
+            transaction_no=500,
+            transaction_type=1,
+            generate_date_time="2024-06-01T12:00:00Z",
+        )
+
+        with pytest.raises(CannotCreateException):
+            await repo.create_tranlog_async(tranlog)
+
+    @pytest.mark.asyncio
+    async def test_get_tranlog_by_transaction_no_builds_correct_query(self):
+        repo = self._make_repo()
+        repo.get_one_async = AsyncMock(return_value=None)
+
+        await repo.get_tranlog_by_transaction_no_async("S001", 1, 100)
+
+        query = repo.get_one_async.call_args[0][0]
+        assert query == {
+            "tenant_id": "T001",
+            "store_code": "S001",
+            "terminal_no": 1,
+            "transaction_no": 100,
+        }
+
+
+# =========================================================================
+# TerminalCounterRepository tests
+# =========================================================================
+
+
